espresso-backend-services/requirements.txt
attrs==22.1.0
autopep8==2.0.1
click==8.1.3
dnspython==2.2.1
exceptiongroup==1.0.4
Flask==2.2.2
Flask-Login==0.6.2
importlib-metadata==5.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
pluggy==1.0.0
pycodestyle==2.10.0
pymongo==4.3.3
pytest==7.2.0
python-dotenv==0.21.0
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.11.0
